pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentfactory"
version = "0.1.0"
description = "A meta-agent system: agents that create specialized agents with tools."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Christopher Hartline"},
]

dependencies = [
    "langchain-core>=0.3",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langgraph>=0.2",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
# Full runtime with deepagents support
full = [
    "deepagents",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
agentfactory = ["templates/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
